// src/app/api/models/user-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AvatarDto } from '../models/avatar-dto';
import { DailyGameSessionDto } from '../models/daily-game-session-dto';
import { FavoriteGameDto } from '../models/favorite-game-dto';
import { GameProgressDto } from '../models/game-progress-dto';
import { GameRecommendationDto } from '../models/game-recommendation-dto';
import { MessageDto } from '../models/message-dto';
import { NotificationDto } from '../models/notification-dto';
import { UserAchievementDto } from '../models/user-achievement-dto';
import { UserSanctionDto } from '../models/user-sanction-dto';
import { WinStreakDto } from '../models/win-streak-dto';
export interface UserDto {
  achievementsEarned?: Array<UserAchievementDto>;
  avatar?: AvatarDto;
  birthDate?: string;
  country?: string;
  creationDate?: string;
  dailyGameSessions?: Array<DailyGameSessionDto>;
  email?: string;
  favoriteGames?: Array<FavoriteGameDto>;
  id?: number;
  isActive?: boolean;
  isAdmin?: boolean;
  lastLogin?: string;
  lastModifiedDate?: string;
  messagesReceived?: Array<MessageDto>;
  messagesSent?: Array<MessageDto>;
  notifications?: Array<NotificationDto>;
  onlineStatus?: 'ONLINE' | 'OFFLINE' | 'BUSY';
  password?: string;
  points?: number;
  privacySetting?: 'PUBLIC' | 'PRIVATE' | 'FRIENDS_ONLY';
  progressions?: Array<GameProgressDto>;
  recommendationsReceived?: Array<GameRecommendationDto>;
  recommendationsSent?: Array<GameRecommendationDto>;
  sanctionsReceived?: Array<UserSanctionDto>;
  themePreference?: 'LIGHT' | 'DARK';
  totalGamesPlayed?: number;
  totalPlayTime?: number;
  userRank?: 'BRONZE' | 'SILVER' | 'GOLD' | 'MONSTER';
  username?: string;
  winStreaks?: Array<WinStreakDto>;
}