// src/app/api/models/message-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { UserDto } from '../models/user-dto';
export interface MessageDto {
  content?: string;
  creationDate?: string;
  id?: number;
  isRead?: boolean;
  lastModifiedDate?: string;
  receiver?: UserDto;
  sender?: UserDto;
}